Decouple logical decoding and AVRO format.
Provides for different output formats that can be chosen by the
connecting client using plugin options in the start command:

  START_REPLICATION SLOT <slot> LOGICAL <XXX/XXX> (FORMAT '<format>')

If not specified, AVRO is used as the default output format.
