+++ ext/format-avro.h
@@ -1,0 +1,8 @@
+#ifndef FORMAT_AVRO_H
+#define FORMAT_AVRO_H
+
+#include "replication/output_plugin.h"
+
+void output_format_avro_init(OutputPluginCallbacks *cb);
+
+#endif /* FORMAT_AVRO_H */
