+++ ext/logdecoder.c
@@ -1,170 +1,122 @@
-#include "io_util.h"
-#include "protocol_server.h"
-#include "oid2avro.h"
-
-#include "replication/logical.h"
-#include "replication/output_plugin.h"
-#include "utils/memutils.h"
+#include "logdecoder.h"
+#include "format-avro.h"
+#include "nodes/parsenodes.h"
+#include "utils/elog.h"
 
 /* Entry point when Postgres loads the plugin */
 extern void _PG_init(void);
 extern void _PG_output_plugin_init(OutputPluginCallbacks *cb);
 
-static void output_avro_startup(LogicalDecodingContext *ctx, OutputPluginOptions *opt, bool is_init);
-static void output_avro_shutdown(LogicalDecodingContext *ctx);
-static void output_avro_begin_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn);
-static void output_avro_commit_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn, XLogRecPtr commit_lsn);
-static void output_avro_change(LogicalDecodingContext *ctx, ReorderBufferTXN *txn, Relation rel, ReorderBufferChange *change);
-
-typedef struct {
-    MemoryContext memctx; /* reset after every change event, to prevent leaks */
-    avro_schema_t frame_schema;
-    avro_value_iface_t *frame_iface;
-    avro_value_t frame_value;
-    schema_cache_t schema_cache;
-} plugin_state;
-
-void reset_frame(plugin_state *state);
-int write_frame(LogicalDecodingContext *ctx, plugin_state *state);
+static void output_startup(LogicalDecodingContext *ctx, OutputPluginOptions *opt, bool is_init);
+static void output_shutdown(LogicalDecodingContext *ctx);
+static void output_begin_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn);
+static void output_commit_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn, XLogRecPtr commit_lsn);
+static void output_change(LogicalDecodingContext *ctx, ReorderBufferTXN *txn, Relation rel, ReorderBufferChange *change);
 
 
 void _PG_init() {
 }
 
 void _PG_output_plugin_init(OutputPluginCallbacks *cb) {
+    elog(DEBUG1, "bottledwater: _PG_output_plugin_init");
     AssertVariableIsOfType(&_PG_output_plugin_init, LogicalOutputPluginInit);
-    cb->startup_cb = output_avro_startup;
-    cb->begin_cb = output_avro_begin_txn;
-    cb->change_cb = output_avro_change;
-    cb->commit_cb = output_avro_commit_txn;
-    cb->shutdown_cb = output_avro_shutdown;
+    cb->startup_cb = output_startup;
+    cb->begin_cb = output_begin_txn;
+    cb->change_cb = output_change;
+    cb->commit_cb = output_commit_txn;
+    cb->shutdown_cb = output_shutdown;
 }
 
-static void output_avro_startup(LogicalDecodingContext *ctx, OutputPluginOptions *opt,
+static void output_startup(LogicalDecodingContext *ctx, OutputPluginOptions *opt,
         bool is_init) {
-    plugin_state *state = palloc(sizeof(plugin_state));
-    ctx->output_plugin_private = state;
-    opt->output_type = OUTPUT_PLUGIN_BINARY_OUTPUT;
+    plugin_state *state;
 
-    state->memctx = AllocSetContextCreate(ctx->context, "Avro decoder context",
+    elog(DEBUG1, "bottledwater: output_startup: is_init=%s", is_init ? "true" : "false");
+
+    state = palloc0(sizeof(plugin_state));
+    ctx->output_plugin_private = state;
+
+    state->memctx = AllocSetContextCreate(ctx->context, "Bottledwater decoder context",
             ALLOCSET_DEFAULT_MINSIZE, ALLOCSET_DEFAULT_INITSIZE, ALLOCSET_DEFAULT_MAXSIZE);
 
-    state->frame_schema = schema_for_frame();
-    state->frame_iface = avro_generic_class_from_schema(state->frame_schema);
-    avro_generic_value_new(state->frame_iface, &state->frame_value);
-    state->schema_cache = schema_cache_new(ctx->context);
+    if (ctx->output_plugin_options) {
+        ListCell *o;
+
+        foreach(o, ctx->output_plugin_options) {
+          DefElem *e;
+
+          e = (DefElem *) lfirst(o);
+          if (strcasecmp(e->defname, "FORMAT") == 0) {
+              char *str;
+
+              if (e->arg) {
+                str = ((Value *) e->arg)->val.str;
+              } else {
+                  ereport(ERROR,
+                          (errcode(ERRCODE_UNDEFINED_PARAMETER),
+                           errmsg("FORMAT option requires an argument")));
+              }
+              if (state->format_cb) {
+                  ereport(ERROR,
+                          (errcode(ERRCODE_AMBIGUOUS_PARAMETER), // TODO: a better one?
+                           errmsg("multiple FORMAT options specified for output plugin")));
+              }
+              if (strcasecmp(str, "AVRO") == 0) {
+                  ; // default format, see below
+              } else {
+                  ereport(ERROR,
+                          (errcode(ERRCODE_FEATURE_NOT_SUPPORTED),
+                           errmsg("unsupported FORMAT option: %s", str)));
+              }
+          }
+        }
+    }
+
+    // use default format, if not created earlier
+    if (!state->format_cb) {
+      state->format_cb = palloc0(sizeof(OutputPluginCallbacks));
+      output_format_avro_init(state->format_cb);
+    }
+
+    state->format_cb->startup_cb(ctx, opt, is_init);
 }
 
-static void output_avro_shutdown(LogicalDecodingContext *ctx) {
+static void output_shutdown(LogicalDecodingContext *ctx) {
     plugin_state *state = ctx->output_plugin_private;
-    MemoryContextDelete(state->memctx);
 
-    schema_cache_free(state->schema_cache);
-    avro_value_decref(&state->frame_value);
-    avro_value_iface_decref(state->frame_iface);
-    avro_schema_decref(state->frame_schema);
+    state->format_cb->shutdown_cb(ctx);
+
+    MemoryContextDelete(state->memctx);
 }
 
-static void output_avro_begin_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn) {
+static void output_begin_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn) {
     plugin_state *state = ctx->output_plugin_private;
     MemoryContext oldctx = MemoryContextSwitchTo(state->memctx);
-    reset_frame(state);
 
-    if (update_frame_with_begin_txn(&state->frame_value, txn)) {
-        elog(ERROR, "output_avro_begin_txn: Avro conversion failed: %s", avro_strerror());
-    }
-    if (write_frame(ctx, state)) {
-        elog(ERROR, "output_avro_begin_txn: writing Avro binary failed: %s", avro_strerror());
-    }
+    state->format_cb->begin_cb(ctx, txn);
 
     MemoryContextSwitchTo(oldctx);
     MemoryContextReset(state->memctx);
 }
 
-static void output_avro_commit_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn,
+static void output_commit_txn(LogicalDecodingContext *ctx, ReorderBufferTXN *txn,
         XLogRecPtr commit_lsn) {
     plugin_state *state = ctx->output_plugin_private;
     MemoryContext oldctx = MemoryContextSwitchTo(state->memctx);
-    reset_frame(state);
 
-    if (update_frame_with_commit_txn(&state->frame_value, txn, commit_lsn)) {
-        elog(ERROR, "output_avro_commit_txn: Avro conversion failed: %s", avro_strerror());
-    }
-    if (write_frame(ctx, state)) {
-        elog(ERROR, "output_avro_commit_txn: writing Avro binary failed: %s", avro_strerror());
-    }
+    state->format_cb->commit_cb(ctx, txn, commit_lsn);
 
     MemoryContextSwitchTo(oldctx);
     MemoryContextReset(state->memctx);
 }
 
-static void output_avro_change(LogicalDecodingContext *ctx, ReorderBufferTXN *txn,
+static void output_change(LogicalDecodingContext *ctx, ReorderBufferTXN *txn,
         Relation rel, ReorderBufferChange *change) {
-    int err = 0;
-    HeapTuple oldtuple = NULL, newtuple = NULL;
     plugin_state *state = ctx->output_plugin_private;
     MemoryContext oldctx = MemoryContextSwitchTo(state->memctx);
-    reset_frame(state);
 
-    switch (change->action) {
-        case REORDER_BUFFER_CHANGE_INSERT:
-            if (!change->data.tp.newtuple) {
-                elog(ERROR, "output_avro_change: insert action without a tuple");
-            }
-            newtuple = &change->data.tp.newtuple->tuple;
-            err = update_frame_with_insert(&state->frame_value, state->schema_cache, rel,
-                    RelationGetDescr(rel), newtuple);
-            break;
-
-        case REORDER_BUFFER_CHANGE_UPDATE:
-            if (!change->data.tp.newtuple) {
-                elog(ERROR, "output_avro_change: update action without a tuple");
-            }
-            if (change->data.tp.oldtuple) {
-                oldtuple = &change->data.tp.oldtuple->tuple;
-            }
-            newtuple = &change->data.tp.newtuple->tuple;
-            err = update_frame_with_update(&state->frame_value, state->schema_cache, rel, oldtuple, newtuple);
-            break;
-
-        case REORDER_BUFFER_CHANGE_DELETE:
-            if (change->data.tp.oldtuple) {
-                oldtuple = &change->data.tp.oldtuple->tuple;
-            }
-            err = update_frame_with_delete(&state->frame_value, state->schema_cache, rel, oldtuple);
-            break;
-
-        default:
-            elog(ERROR, "output_avro_change: unknown change action %d", change->action);
-    }
-
-    if (err) {
-        elog(ERROR, "output_avro_change: row conversion failed: %s", avro_strerror());
-    }
-    if (write_frame(ctx, state)) {
-        elog(ERROR, "output_avro_change: writing Avro binary failed: %s", avro_strerror());
-    }
+    state->format_cb->change_cb(ctx, txn, rel, change);
 
     MemoryContextSwitchTo(oldctx);
     MemoryContextReset(state->memctx);
-}
-
-void reset_frame(plugin_state *state) {
-    if (avro_value_reset(&state->frame_value)) {
-        elog(ERROR, "Avro value reset failed: %s", avro_strerror());
-    }
-}
-
-int write_frame(LogicalDecodingContext *ctx, plugin_state *state) {
-    int err = 0;
-    bytea *output = NULL;
-
-    check(err, try_writing(&output, &write_avro_binary, &state->frame_value));
-
-    OutputPluginPrepareWrite(ctx, true);
-    appendBinaryStringInfo(ctx->out, VARDATA(output), VARSIZE(output) - VARHDRSZ);
-    OutputPluginWrite(ctx, true);
-
-    pfree(output);
-    return err;
 }
